pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sarohy <a.h.sarohy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"
pydantic = {extras = ["email"], version = "^2.9.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
